VR-175: moving files.
