+++ src/views/queries/responses/__tests__/partial-query.test.ts
