+++ src/views/queries/responses/__tests__/scope3.test.ts
@@ -1,7 +1,7 @@
 import { expect } from 'chai'
 import { describe } from 'mocha'
-import Scope3CarbonConsumptionResponseTemplates from '../../queryResponses/scope3.js'
-import Scope3CarbonConsumptionViewResponseTemplates from '../../queryResponses/viewResponseToScope3.js'
+import Scope3CarbonConsumptionResponseTemplates from '../scope3.js'
+import Scope3CarbonConsumptionViewResponseTemplates from '../viewResponseToScope3.js'
 
 describe('QueryResponseTemplates', () => {
   describe('newScope3CarbonConsumptionResponseFormPage Tests', () => {
