+++ src/views/queries/responses/scope3.tsx
