+++ src/views/queries/responses/viewResponseToScope3.tsx
